Students
Fixed bugs and allowed students to save into database. Previously there was an "address" column in the table, removed it since it has no reference anywhere else in the project, likely a mistake. Restored functionality for creating and updating by allowing status and branch_id to be sent through the API